Revert "change file structure"
This reverts commit a9dc7b342d53be4794071ba2e1b122bbb0654c98.
