+++ module/move/gspread/src/gcore.rs
@@ -6,6 +6,4 @@ crate::mod_interface!
   layer client;
   layer error;
   layer secret;
-  layer types;
-  layer methods;
 }
