--- module/move/gspread/src/gcore/methods.rs
@@ -1,8 +1,0 @@
-
-mod private {}
-
-crate::mod_interface!
-{
-  layer sheet;
-  layer values;
-}
